requirements.txt
alembic==1.6.5
aniso8601==9.0.1
click==8.0.1    
colorama==0.4.4 
Flask==2.0.1    
Flask-JWT==0.3.2
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==1.4.2
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1